#802 move calendar examples to util examples file
