+++ src/a07_timeline_logic/test/_util/calendar_examples.py
@@ -233,3 +233,123 @@ def display_creg_five_squirt_time_attrs(graphics_bool: bool):
         x1_scatter = plotly_Scatter(x=xp_list, y=yp_list, text=curr_list, mode="text")
         x_fig.add_trace(x1_scatter)
         conditional_fig_show(x_fig, graphics_bool)
+
+
+def get_expected_creg_2024_markdown() -> str:
+    return """
+                               Year 2024                                
+
+       March                     April                      May         
+Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su
+             1  2  3       1  2  3  4  5  6  7             1  2  3  4  5
+ 4  5  6  7  8  9 10       8  9 10 11 12 13 14       6  7  8  9 10 11 12
+11 12 13 14 15 16 17      15 16 17 18 19 20 21      13 14 15 16 17 18 19
+18 19 20 21 22 23 24      22 23 24 25 26 27 28      20 21 22 23 24 25 26
+25 26 27 28 29 30 31      29 30                     27 28 29 30 31      
+
+        June                      July                     August       
+Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su
+                1  2       1  2  3  4  5  6  7                1  2  3  4
+ 3  4  5  6  7  8  9       8  9 10 11 12 13 14       5  6  7  8  9 10 11
+10 11 12 13 14 15 16      15 16 17 18 19 20 21      12 13 14 15 16 17 18
+17 18 19 20 21 22 23      22 23 24 25 26 27 28      19 20 21 22 23 24 25
+24 25 26 27 28 29 30      29 30 31                  26 27 28 29 30 31   
+
+     September                  October                   November      
+Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su
+                   1          1  2  3  4  5  6                   1  2  3
+ 2  3  4  5  6  7  8       7  8  9 10 11 12 13       4  5  6  7  8  9 10
+ 9 10 11 12 13 14 15      14 15 16 17 18 19 20      11 12 13 14 15 16 17
+16 17 18 19 20 21 22      21 22 23 24 25 26 27      18 19 20 21 22 23 24
+23 24 25 26 27 28 29      28 29 30 31               25 26 27 28 29 30   
+30                                                                      
+
+      December               January (2025)           February (2025)   
+Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su      Mo Tu We Th Fr Sa Su
+                   1             1  2  3  4  5                      1  2
+ 2  3  4  5  6  7  8       6  7  8  9 10 11 12       3  4  5  6  7  8  9
+ 9 10 11 12 13 14 15      13 14 15 16 17 18 19      10 11 12 13 14 15 16
+16 17 18 19 20 21 22      20 21 22 23 24 25 26      17 18 19 20 21 22 23
+23 24 25 26 27 28 29      27 28 29 30 31            24 25 26 27 28      
+30 31                                                                   """
+
+
+def get_expected_creg_year0_markdown() -> str:
+    return """
+                                 Year 0                                 
+
+       March                     April                      May         
+We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu
+ 1  2  3  4  5  6  7                1  2  3  4                      1  2
+ 8  9 10 11 12 13 14       5  6  7  8  9 10 11       3  4  5  6  7  8  9
+15 16 17 18 19 20 21      12 13 14 15 16 17 18      10 11 12 13 14 15 16
+22 23 24 25 26 27 28      19 20 21 22 23 24 25      17 18 19 20 21 22 23
+29 30 31                  26 27 28 29 30            24 25 26 27 28 29 30
+                                                    31                  
+
+        June                      July                     August       
+We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu
+    1  2  3  4  5  6                1  2  3  4                         1
+ 7  8  9 10 11 12 13       5  6  7  8  9 10 11       2  3  4  5  6  7  8
+14 15 16 17 18 19 20      12 13 14 15 16 17 18       9 10 11 12 13 14 15
+21 22 23 24 25 26 27      19 20 21 22 23 24 25      16 17 18 19 20 21 22
+28 29 30                  26 27 28 29 30 31         23 24 25 26 27 28 29
+                                                    30 31               
+
+     September                  October                   November      
+We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu
+       1  2  3  4  5                   1  2  3       1  2  3  4  5  6  7
+ 6  7  8  9 10 11 12       4  5  6  7  8  9 10       8  9 10 11 12 13 14
+13 14 15 16 17 18 19      11 12 13 14 15 16 17      15 16 17 18 19 20 21
+20 21 22 23 24 25 26      18 19 20 21 22 23 24      22 23 24 25 26 27 28
+27 28 29 30               25 26 27 28 29 30 31      29 30               
+
+      December                January (1)               February (1)    
+We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu      We Th Fr Sa Su Mo Tu
+       1  2  3  4  5                      1  2          1  2  3  4  5  6
+ 6  7  8  9 10 11 12       3  4  5  6  7  8  9       7  8  9 10 11 12 13
+13 14 15 16 17 18 19      10 11 12 13 14 15 16      14 15 16 17 18 19 20
+20 21 22 23 24 25 26      17 18 19 20 21 22 23      21 22 23 24 25 26 27
+27 28 29 30 31            24 25 26 27 28 29 30      28                  
+                          31                                            """
+
+
+def get_expected_five_5524_markdown() -> str:
+    return """
+                                Year 5224                                 
+
+   Fredrick              Geo               Holocene             Iguana    
+An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea
+       0  1  2             0  1  2             0  1  2             0  1  2
+ 3  4  5  6  7       3  4  5  6  7       3  4  5  6  7       3  4  5  6  7
+ 8  9 10 11 12       8  9 10 11 12       8  9 10 11 12       8  9 10 11 12
+13 14 15 16 17      13 14 15 16 17      13 14 15 16 17      13 14 15 16 17
+18 19 20 21 22      18 19 20 21 22      18 19 20 21 22      18 19 20 21 22
+23 24               23 24               23 24               23 24         
+
+    Jesus                Keel               LeBron             Mikayla    
+An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea
+       0  1  2             0  1  2             0  1  2             0  1  2
+ 3  4  5  6  7       3  4  5  6  7       3  4  5  6  7       3  4  5  6  7
+ 8  9 10 11 12       8  9 10 11 12       8  9 10 11 12       8  9 10 11 12
+13 14 15 16 17      13 14 15 16 17      13 14 15 16 17      13 14 15 16 17
+18 19 20 21 22      18 19 20 21 22      18 19 20 21 22      18 19 20 21 22
+23 24               23 24               23 24               23 24         
+
+    Ninon               Obama              Preston              Quorum    
+An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea
+       0  1  2             0  1  2             0  1  2             0  1  2
+ 3  4  5  6  7       3  4  5  6  7       3  4  5  6  7       3  4  5  6  7
+ 8  9 10 11 12       8  9 10 11 12       8  9 10 11 12       8  9 10 11 12
+13 14 15 16 17      13 14 15 16 17      13 14 15 16 17      13 14 15 16 17
+18 19 20 21 22      18 19 20 21 22      18 19 20 21 22      18 19 20 21 22
+23 24               23 24               23 24               23 24         
+
+  RioGrande             Simon               Trump     
+An Ba Ch Da Ea      An Ba Ch Da Ea      An Ba Ch Da Ea
+       0  1  2             0  1  2             0  1  2
+ 3  4  5  6  7       3  4  5  6  7       3  4  5  6  7
+ 8  9 10 11 12       8  9 10 11 12       8  9 10 11 12
+13 14 15 16 17      13 14 15 16 17      13 14         
+18 19 20 21 22      18 19 20 21 22                    
+23 24               23 24                             """
